feat: move steps to tests examples path
